Added more JavaDoc comments.
I also added another method to remove a task, so that it can be accomplished by either passing in a Task! object, or simply providing the Task! object's unique ID.  We may not want both methods, but for now this gives us some flexibility.  I also added another constructor for a Task object that only takes the title parameter.  This means that we'll probably want to set some reasonable defaults for the other parameters.