Fixed a display bug for a completed task
After clicking on a tast to complete it, the task wasn't displaying as completed.  This was due to two complete sets of views (list and card) being kept in the same TreeMap collection using the same keys.  I fixed this by having a TreeMap for each view type.
Added code to set the app's title via the WebManager
Added code to set the image when the app is saved to a smartphone's homescreen
Updated the onTaskUpdate() method to consolidate the code based on the two view types
Simplified the removeTask() method that took a Task as a paremeter