Added the ability to export/import the tasks to/from a client file
Added Alex's name to the credits
Removed some of the field variables that didn't need to be global, including the preferences and the FAB button
Changed the onImport() and onExport() to be importTasks() and exportTasks(), which deal with importing/exporting the tasks to/from browser storage
Now optionally set the tooltip text on the toolbar buttons based on a preference
Changed the icon and the functionality of the import/export buttons on the toolbar to import/export to a JSON client file
Changed some of the e! event variables to be named event!
Added a method to removeAllTasks(), which must be done before importing tasks from a saved file
Changed the app to only create 5 sample tasks instead of 50